fix: remove security module
Move all functions to the security API module and move
all associated tests to other related test files
