+++ uaclient/api/u/pro/security/fix/cve/execute/v1.py
@@ -2,6 +2,7 @@ from typing import List
 
 from uaclient.api.api import APIEndpoint
 from uaclient.api.data_types import AdditionalInfo
+from uaclient.api.u.pro.security.fix import FixStatus
 from uaclient.api.u.pro.security.fix._common import get_expected_overall_status
 from uaclient.api.u.pro.security.fix._common.execute.v1 import (
     FixExecuteResult,
@@ -13,7 +14,6 @@ from uaclient.api.u.pro.security.fix.cve.plan.v1 import (
 )
 from uaclient.config import UAConfig
 from uaclient.data_types import DataObject, Field, StringDataValue, data_list
-from uaclient.security import FixStatus
 
 
 class CVEFixExecuteOptions(DataObject):
